Add `PatientArchiver`
This adds a service class which is responsible for the steps that need
to be taken when archiving a patient. Currently this logic is in the
`PatientArchiveForm`, but by moving it in to a separate service class we
can unit test it separately and in the future integrate it in to other
parts of the service (most likely ops tools).

Jira-Issue: MAV-1918