Add index on `pending_changes <> '{}'`
This `WHERE` clause is used on each page to render the count of import
issues in the navigation bar, therefore we should try and speed up this
query as it will have an impact on every page render. Adding an index on
this filter should ensure that selecting the number of patients or
vaccination records with import issues is quicker.

Jira-Issue: MAV-2001