Standardise imports and offline sheet to accept Imms API records
There are a number of issues which cause unedited Imms API records
to flag as duplicates when uploading an offline spreadsheet. These
are listed below.

## Round time objects when comparing in `stage_changes`

There was an issue with detecting import duplicates on
`VaccinationRecord`s; the value in the offline spreadsheet was
rounded to the nearest second, whereas the value in the database
was accurate to the nearest millisecond. This was causing the
times to show as a duplicate to review unnecessarily.

## ODS code

The offline spreadsheet used to always use the session's ODS code,
not the existing `VaccinationRecord`'s.

## Performer's names

I see no reason to overwrite these if they're `nil` in the upload.

## Protocol

Should always be `nil` for an Imms API record.

## Location

This can't be properly encoded in the offline spreadsheet, so instead
have ignored it on import if importing an Imms API record.