Discard `ConsentFormMatchingJob` once matched
This ensures that if a consent form has already been matched, the
automated process doesn't run and match consents with patients which
might result in duplicate consents.

This is necessary as we've moved this job to Sidekiq where it will
continue to retry if it fails. For example, it might fail due to a bad
request from PDS, and continue to do this for the next 21 days. In that
time, the consent form may have been matched manually and therefore we
would never want the automatic job to succeed.