Handle invalid phone errors from GOV.UK Notify
This ensures that we handle the `InvalidPhoneError: Not a UK mobile
number` error from GOV.UK Notify by recording the response as a
permanent failure (which is accurate since retrying is not going to
succeed). This also means the job finishes and doesn't get put back on
the retry queue, and also allows the nurses to see useful feedback.

Sentry-Issue: 6235087330