Add task to mark vaccinations as synced to NHS
These vaccinations were synced to the NHS Imms API, however they
triggered an error when saving the sync information to the db. The
underlying issue has been fixed and we've extracted the Imms API
identifier, and are adding them to the records with this data migration.

This migration can be removed once performed in production.

Jira-Issue: MAV-2018