Don't delete teams in a transaction
This endpoint is only used for testing and having it operate in a
transaction makes it less useful for testing. It means that if we get a
timeout, the whole action has failed, whereas the testers would prefer
to be able to simply run it again to delete everything that was left
behind. It should also improve performance and remove the chance of
deadlocks.