Specify `NOT VALID` when adding foreign keys
This updates two migrations which currently take a very long time to run
to skip the validation step when adding a foreign key. We can do this
because we're modifying existing foreign keys so we know they will
already be valid, and this will massively improve the performance of
these migrations.