Add academic years to `LocationProgrammeYearGroup`
This adds a new column, `academic_year`, to the
`location_programme_year_groups` table. This allows us to have a
different set of year groups at a particular location for a particular
programme for each academic year. This is necessary because the table
currently holds the value for the current academic year, but if this
were to change then it would affect which patients are shown in
sessions in previous academic years.

For example, if flu is given to year 12 and 13 at a school one year,
and then next year it's no longer given to those year groups, any
sessions in the previous academic year will lose visibility of the
year 12 and 13 patients.

Jira-Issue: MAV-2044