Assert that Imms API records' `Batch.team` is `nil`
Without this, the batch can match on `expiry`, `name` and `vaccine`,
even if this batch appears in a team. This causes issues when uploading
an offline spreadsheet, because that assumes that all API records'
`Batch.team` values are `nil`.

This was seen in testing, and is not expected to be seen very often,
unless somehow Mavis ends up consuming records which were created
by Mavis in the API. We don't expect this to happen.