Gracefully handle page overflow errors
This configures Pagy to ensure that if a user tries to access a page
that doesn't exist, rather than seeing an error page (and an unhandled
exception going to Sentry), the user is shown the last possible page.

To support this I've also reconfigured how we include the Pagy extras to
make it globally configured across the codebase. This is because the
Pagy behaviour was dependent on which files had been executed and which
`pagy/extra` files had been required, meaning the tests were flaky
depending on the order.