Ensure patients are visible when PSDs only are enabled
This fixes a bug where patients don't appear to show in the "Record
vaccinations" tab if they've got a PSD and only PSDs are enabled for the
session (i.e. national protocol is switched off).

The bug was happening because we were not handling the situation where
only PSD is enabled for the session, in that case the scope was being
reduce to find no patients.

Jira-Issue: MAV-2065