Model patient session eligibility
This adds a scope which makes it possible to filter on patients in a
session who are "eligible" for the vaccination. We're not sure about the
naming of this term, and we might change it later, but for now the logic
is any patients who haven't been vaccinated elsewhere or weren't already
vaccinated last year.

Jira-Issue: MAV-2061