Ensure status matches academic year
This fixes an issue where the status of a patient in a particular
academic year does not reflect their status in that academic year. For
example, a vaccination status for the 2024/25 academic year, will
currently show as vaccinated if the patient receives the vaccination in
the 2025/26 academic year. This means the status cannot currently be
used to determine whether a patient is "eligible" for a vaccine in any
given academic year.

Jira-Issue: MAV-2061