Ensure `Patient#teams` doesn't return duplicates
If a patient belongs to multiple sessions for the same team (which is
normal) then this method returns the same team multiple times. Although
this is not really an issue in the way this method is used, it does
cause a problem when creating archive reasons when the patient is
deceased as we try and create multiple archive reasons for the same team
which results in a unique constraint failing.

Sentry-Issue: 6902363114