Fix `ConsentForm#session`
This fixes a couple of issues with the method to get the session for the
consent form.

Firstly this fixes a regression that was added in
c5a5879ea866b31786bdc21e595bd7d8ee56136d where if a parent tells us that
their child is actually home schooled, the session remains the school
session meaning there's a chance they will see the deadline error
message if the school session has no more dates (when actually they
should fill in consent for the clinic session).

Secondly this fixes an issue that was only partially fixed in
c5a5879ea866b31786bdc21e595bd7d8ee56136d where we are trying to pick the
most approriate session if there are numerous possible candidates (i.e.
multiple sessions at the same location with the same programmes). In the
original change we prioritised scheduled session over completed
sessions, but not unscheduled ones meaning there was still the chance
for the wrong session to be selected.

I've also tried to improve test coverage to avoid regressions in the
future.

Jira-Issue: MAV-2036