Update invalidated patients from PDS
This should improve dealing with cases where a patient was invalidated
without an NHS number, but they now have an NHS number.

There are currently 75 patients in production marked as invalid. We
first want to check whether any of these are now superseded, and
whether, if we were to refresh their details from PDS, we would get new
patient details for them. If this is the case then we should ensure
that invalidated patients are also updated from PDS as part of the
nightly background job.

Jira-Issue: MAV-1962