Validate `year_group` column
This adds validation to the `LocationProgrammeYearGroup` model ensuring
that the year group is suitable for the location. This should avoid a
situation where a programme year group is added to a location, but the
location doesn't include that year group meaning the patients don't
appear correctly in sessions.

Jira-Issue: MAV-2136