Allow submitting consent for an unscheduled session
When filling in a consent form and recording that the child goes to a
different school to the one we thought, this ensures that the consent
form doesn't show the "deadline passed" end state if the new school
session is unscheduled. This situation could occur if the new school
doesn't have dates yet, but will do in the future, meaning that
submitting consent at this point is worth doing.

This partially reverts 63b0d9329aef8db8746201a27494703ab458ec65 which
introduce this regression in this specific scenario.

Jira-Issue: MAV-2144