Add `gias_year_groups` column
This adds a new column to the `locations` table that will behave the
same as the current `year_groups` column, to clarify it's name and where
the underyling data comes from. To support a zero downtime migration,
this adds the new column and starts using it. The old column will be
removed in a future commit.

Jira-Issue: MAV-2148