Configure Redis cache store
This sets up the Rails cache to use the shared Redis cache to allow for
caches to be shared between the application instances. In theory this
might slightly improve performance out of the box (only one machine
needs to cache something before it can be used by all the others),
however we're not currently making extensive use of Rails caching so the
real performance gains are likely to come later when we start doing
that.