Return totals from reporting API
This creates a PatientProgrammeStatus model that's backed by a
materialised view, implemented using the
[scenic](https://github.com/scenic-views/scenic) gem from Thoughtbot.

A view in Postgres is a table that runs a query every time it's
requested.

A materialised view is a table that's precomputed before time,
effectively like a cache.

Materialised views are refreshed in either a non-concurrent or
concurrent fashion. A non-concurrent refresh locks the table (can't be
read), whereas a concurrent refresh allows the data to be queried even
while it's being computed.

Refreshes can be scheduled to run using a background job, at certain
times, or as a result of callbacks. For this PR, we haven't implemented
those. Instead we can simply run
`ReportingAPI::PatientProgrammeStatus.refresh!` from a console or Rails
runner to test for now.

The scenic generator for views creates a `db/views/file_v01.sql` which
is the SQL definition of the view, plus a migration using a
`change_view` call that stands up the view using the specified SQL, plus
can be populated with a few indexes.

The SQL file is meant to be written as one big chunky SQL, joining any
data it needs. It has an ID column that's concatenated the various data
sources it's pulling from, for debugging, but also for ensuring the
`concurrent` refresh works correctly (as that requires a unique key).

When we want to change the structure of a view, we need to create a
`v02.sql` which replaces it. `scenic` smartly figures out if we're
replacing a view, and adjusts the migration accordingly.

The materialised view approach to creating queryable caches of reporting
data is different from the existing VaccinationEvent pattern, which is a
standard ActiveRecord-backed model, that denormalises data from multiple
other models.

The `VaccinationEvent` pattern we are using elsewhere in the
ReportingAPI namespace is accomplishing the same as the view, but in
Ruby.

The main advantage of materialised views is their performance. Not
having to deal with ActiveRecord marshalling, serialisation, and other
bits means that they should run very fast even with millions of records.

With the Ruby version we could implement some form of incremental
updating, which we can't do with the view; you have to recreate the data
wholesale every time you run `.refresh!`. However, in practice,
recreating the data inside SQL is likely to be orders of magnitudes
faster than any Ruby operation.

Views are a bit more annoying to write, you have to write raw SQL, not
pretty Ruby.

But overall, the complexity of creating multiple AR-backed cache-type
models, migrating those, keeping everything in sync, balances out versus
the "just blow away the view, rebuild it fresh every time we want a new
column" simplicity of views.

Only the happy path is tested at the controller level because we will
likely iterate and churn this endpoint and view a fair bit, so
exhaustive testing isn't worth the investment at this point.