Fix inaccurate access logging in timeline ops tool
The access log entries were being recorded as an after_action. In conjunction with other logic which redirected with default parameters if some parameters were unset, this led to extra log entries being recorded, with incorrect details.

This has been fixed by adjusting the sequencing of the access log recording, putting it inside the controller.