Add `contains_gelatine` to vaccines
At the moment to determine whether a vaccine contains porcine gelatine
or not we check whether the programme is flu and the method is nasal
spray. This works as the only vaccine that contains gelatine and is
approved for flu is the nasal spray. However, for MMR there are two
kinds of vaccines that are both injection, one contains gelatine and one
does not, meaning we need to store this in the table itself.

Jira-Issue: MAV-2069