Fix flaky tests related to MMR being added
The MMR programme is catch up only meaning that it doesn't create any
sessions by itself. The criteria for when a patient is considered
vaccinated is also a little different to the other programmes as
patients required two doses.

This means any tests which rely on the behaviour of non-MMR programmes,
but pick a random programme are liable to be flaky.