Stop using columns to be removed
This adds the `location_id`, `academic_year` and `year_group` columns
from the `Locaion::ProgrammeYearGroup` model since they have been
replaced with the association to the `Location::YearGroup` model.

The columns will be removed in a future commit, but this will be
deployed first to ensure the columns can be removed without being used.

Jira-Issue: MAV-2148