Export separate CarePlus row per session date
This fixes an issue where if a patient is vaccinated multiple times in
the same session, but on differen dates, then only one row is generated
in the report. This is not relevant for any of the non-MMR programmes as
patients would only be vaccinated once, but for MMR it's possible that
patient will be vaccinated twice in the same session.

Jira-Issue: MAV-1248