Fix access logging of visible fields for inspect graph
The access logging for visible fields in the inspect graph page was not working properly.

The old approach traversed the graph and used the relationships between nodes to decide which types of objects were visible.
This caused problems when relationships were via an intermediary object (for example "User" is linked to "VaccinationRecord" via the "performed_by_user" attribute.

This has been fixed by adding a function in the graph renderer to return a list of node types rendered in the graph.

Jira-Issue: MAV-2249