Add `without_gelatine` columns
This adds a new column to the `ConsentFormProgramme` model which will
capture whether a parent requests that their child doesn't receive any
vaccine with gelatine in.

This also adds a new column to the `Consent` model which will be used
for capturing whether a parent asks their child not to receive the
vaccine with gelatine.

Jira-Issue: MAV-2074