Fix access logging of visible fields for inspect timeline
Previous logic led to incorrect access log entries being saved, in particular when the page was loaded for the first time or when nothing was selected on the left-hand side.

This PR changes two things.
Firstly, it tightens the conditions under which default values are set, so that it should only happen when the page is loaded for the first time.
Secondly, it tidies up how the set of events visible on the left-hand side are managed.

In the medium term this code should be refactored to merge `event_names` and `detail_config`.

Jira-Issue: MAV-2232