Deduplicate Imms API records based on `primarySource`
This deduplicates based on the date of vaccination and the programme.

The deduplication prefers records where `primarySource == true`, but
will accept records where `primarySource == false` if none are `true`.

If multiple records have `primarySource == true`, then they will all
be accepted; we have no way of telling which record is more reliable.

The record's `performed_at` time is rounded to the nearest day for
comparison.

Jira-Issue: MAV-2257