Display PDS search results for NHS numbers found in daily job
Previously, only NHS numbers found from imports were being displayed via the PDS history link on the global child record.
Now that we run the cascading search in the daily job as well, we should display search results if an NHs number is found that way too. This broadens the scope to allow that.

Jira-issue: MAV-2265