Refactor vaccination criteria logic
The logic needs to become more complicated for MMR, but effectively
we're duplicating code in both this class and in the
`StatusGenerator::Vaccination` class so it makes sense to combine the
two in to a single class so all the logic is in one place.

I've also added some comments explaining the logic behind its usage.

Jira-Issue: MAV-2082