Remove "Contraindicated" from tallies
This box previously covered both contraindicated and consent refused,
but it relies on the "Could not vaccinate" status which is shortly going
to be removed. Instead we can remove the contraindicated patients and
show only those without consent. We know that the proportion of
contraindicated patients in any given session is very slow, and this
also fixes an issue where we have been double counting these patients
as they also have a consent status of given.

Jira-Issue: MAV-2137