Refactor how statuses are rendered
This introduces a new class `PatientStatusResolver` which can be used to
determine the final status as it should be shown to the user (passed to
the `AppProgrammeStatusTagsComponent`). This is necessary to consolidate
some of the existing logic around the final status being a combination
of different columns, while also allowing for future work related to
showing either consent, triage or session status for the vaccination
status.

Jira-Issue: MAV-2082