Add new columns to `VaccinationRecord`
This change adds `nhs_immunisations_api_identifier_system`
and `nhs_immunisations_api_identifier_value` to `VaccinationRecord`.

We think these values will be useful if we ever need to deduplicate
records in the future, and it's also quite feasible that we might
want to expose this to users at some point.

Jira-Issue: MAV-2275