Add filter on consent without gelatine
This adds the ability for users to filter patients based on whether they
consented to a vaccine without gelatine or not. This works in the same
way as when users can filter based on whether or not a patient consented
to the nasal spray or injection vaccine for flu.

Jira-Issue: MAV-2425