Add `EnqueuePatientsAgedOutOfSchoolsJob`
This refactors the `PatientsAgedOutOfSchoolJob` to accept a school ID as
a single parameter and then introduces a new job,
`EnqueuePatientsAgedOutOfSchoolsJob`, which then queues the original job
for each school attached to a team. This then allows us to only call the
original job for specific schools when importing a class list or a
cohort avoid unnecessary work and activity in the database.

Jira-Issue: MAV-2420