Keep the order of patients in search batch
Ruby's Set is unordered. We want the patient searches to keep their
order so that the session searches are enqueued before the rolling
searches. This doesn't guarantee that they are performed first, but that
is the intention. To guarantee this we have to make bigger changes to
how we use Sidekiq, so this will have to do until we can do that.

Jira-Issue: MAV-1995