Onboard from GIAS year groups
This fixes an issue where onboarding new training teams or using the
API to find locations to onboard with (as the tests do) would find new
schools because until the schools are added to a team they have no
`Location::YearGroup` instances. Instead we can find locations based on
their GIAS year groups which was the original intention anyway.