Only always show update triage outcome for MMR
This amends 7b47e1470489bab560572ea2b4e17687e57f99b5 to only apply to
MMR to avoid this feature needing to be tested for all programmes and
potentially blocking the next release.

Instead, only for the MMR programme will users be able to update the
triage outcome even if triage is not required.

Jira-Issue: MAV-2303