Use `PatientStatusResolver` in components
This updates the consent, triage and vaccination section components
shown on the patient session page to use the `PatientStatusResolver`
that is used to render the status tags.

By using this class we ensure that the colours and headings will remain
in sync between the statuses shown in a list of patient cards and the
statuses shown on this page when viewing a patient.