Migration for PatientChangeset table
Makes row_number a nullable column so we can create changesets for patients not in the csv file, for simulating school moves out int he review screen.
Also addes record type enum and renames the pending changes columns to data for consistency across other models.

This is to support work on the import review screen.

MAV-906