Fix bug where users cannot search for sessions
This fixes an issue where users cannot search for sessions, fixing a
regression that was added in 60e3b0deb031ea09c9a793b03f334c1577e18039.

In that commit I switched from `pagy_array` to `pagy` thinking that
since the tests pass it was safe to do, however we were missing a test
that performed a search on the page and the query fails to run without
it.

This is an issue in `main` so we'll need to rebase `next` once this is
merged in.