Remove `patients` association on locations
This has the potential for confusion as you might expect it to fetch the
patients through the patient locations but it actually goes via the
`school_id`. This in particular doesn't work for clinics where
`patients` will always be empty since no patients will have a
`school_id` set to a clinic.