Hide dose number row in certain circumstances
For multi-dose programmes like MMR or Td/IPV, we infer the dose sequence
based on the vaccination history and therefore don't set an explicit
dose sequence when recording the vaccination. This results in a row
saying the dose number is unknown which could be potentially confusing
for users. Instead, we can hide the row when the dose sequence is
unknown and the programme doesn't have a default one.