Ensure PSDs are invalidated correctly
This fixes a bug where PSDs (patient specific directions) weren't being
invalidated correctly when a new consent came in for a patient either
via the consent form or via the nurse recording consent. Specifically,
if a patient has consent for nasal flu (and optionally injection) and a
PSD is added allowing a healthcare assistant to perform the vaccination.
However, if a new consent comes in for the patient which refuses nasal
spray, although this would not result in a conflicting consent, we
should invalidate the PSD to make sure a nurse records the vaccination.

Jira-Issue: MAV-2283