Fail import for duplicate NHS number or matching patients
Currently, we fail uploads which have the same NHS number repeated in the file, on import as part of the ProcessImportJob. However this state can also be achieved after the PDS search where two or more rows in the upload have the same NHS number in PDS. In these cases, we currently silently de-dupe and take one of the records only. We should instead fail the upload in these cases too.
It is possible for two rows in the upload, without NHS number, to both match the same patient in the database. We currently arbitrarily choose one of these rows and ignore the other to avoid conflicting import issues staged. We should instead fail the upload in these cases too.

MAV-2476