Return empty navigation items when no team
Also hides the global navigation completely when there are no
navigation items to avoid render empty containers.

This fixes an issue on the select team page where the navigation
would otherwise be shown and where we now can no longer display
the navigation as it is now dependent on the team.