Add method to validate the schools are in the right state now
The warnings for schools without sessions are purposfully excluded from, the 
validation inside the perform as a school ending up without sessions can happen 
if that school did not have sessions to begin with. Rather than resolving that 
in this script we will use the existing Mavis CLI tool to resolve such an issue
if it occurs.

Jira-Issue: MAV-2352
Jira-Issue: MAV-2353