Use `CachedProgramme`
This replaces the use of FactoryBot to create programme instances in the
tests to instead use the new `CachedProgramme` class which should result
in a performance improvement to the tests.

Specifically, using GitHub Actions before this change the tests would
take about 20 minutes to complete, but now complete in around 16
minutes.

Jira-Issue: MAV-2527