Scope overnight status updates to current academic year
This follows on from 7b019fb4f8f2ce5ea9869e752a37597780a25e63 to also
limit the updates that we perform to be based on the academic year.
Previously the change meant that we wouldn't create status records for
other academic years, but now this also reduces the amount of updates
that take place.