Cache navigation counts to a cookie
The Reporting app needs a way to display an identical header to the one
that's in Mavis. The navigation items are easy enough, they're links,
but the tricky bits are the counts.

We considered a few different approaches:

- JSON endpoint to fetch latest counts (works but complex)
- Storing counts in the JWT (works, but cached on login, can go stale)

In the end the simplest solution we think is to use a minimal cookie,
since the two apps live on the same domain.

The cookie is set automatically on every navigation through Mavis, and
the user will be coming to the Reporting app from Mavis. So at the point
of reaching the app, they will always have pretty up to date
information.

Where it can go out of sync is if another user makes a change in Mavis
(matches a consent response), and the user on the Reporting app
refreshes the Reporting dashboard expecting to see the change. They
won't, as they're still using the same cached cookie.

However, if the same user in another tab in Mavis does something, the
next time they interact with the Reporting app, they will see the
updated counts (as the cookie should be refreshed).

One downside of cookies is that they add bloat to every request, but
this cookie should only be a few dozen bytes.