Remove `session` parameter from `StatusUpdater`
There's only one place where we need to update all the patients status
in a session, and therefore we can replace its use with a loop and
enqueue of a job per patient. This is something we want to move towards
anyway.

It also allows us to move towards a world where we can update the status
of a patient even if they don't belong to a particular session.

Jira-Issue: MAV-2520