Update patient statuses regardless of session
This refactors the `StatusUpdater` to ensure that the status of all
patients (include those not currently part of any session) get updated.

By doing this we fix a bug where patients who are not currently part of
any session (but are still visible to a team) would have stale statuses.
An example of how this can happen is if a historical vaccination record
is uploaded, a patient is created and added to the team, but not added
to any sessions.

Jira-Issue: MAV-2520