Run status updates in parallel
This updates the StatusUpdaterJob to allow updating the status of a
single patient in parallel to speed up overall performance of the job.

To support this I've had to convert the job to pure Sidekiq, which
means we should be careful when deploying this to production to make
sure there are no jobs running at the time.