Restore deleted attributes on patient changeset if pending changes
The previous implementation of pending changes from a patient changeset deleted auto-accepted address values from the changeset's child attributes, and did not restore them if they were no-longer considered auto-acceptable changes becasue some other change needs to be reviewed. This approach was acceptable as we only ever called changeset.patient in the commit job and hence the attributes were not relevant anymore and not used again. Now we call changeset.patient in the uniqueness check before the commit job. This could mean that if address details are deleted incorrectly, we might create the wrong patient in the commit job. Hence, just like restoring auto-accepted changes in pending changes, we should restore the attributes on the changeset as well.