Set programme types
This ensures that any `programme_type` or `programme_types` columns are
set at the same time as the existing programme foreign key columns. By
ensuring the two columns remain in sync, this changes remains backwards
compatible with old versions of the code which allows us to deploy this
with zero downtime.

Jira-Issue: MAV-2527