Add per-programme feature flagging for `search`
This will allow us to switch on the `Search` functionality programme
by programme, rather than switching all the remaining programmes on
at the same time. This is particularly useful for the `Search` job,
because we will need to have a "data quality assessment period" for
each of the new programmes. This will allow us to switch the
respective programmes on when we're ready, rather than waiting for
all of the quality assessments to be complete.

This uses actors.

Jira-Issue: MAV-2490