Ignore instruction row in CSV imports
The second row of the class and cohort list templates contains useful information on which fields are optional and which are required, and how fields should be formatted.
Users often experience errors when they leave this row in the CSV which they upload to Mavis.

This change allows Mavis to ignore this row and accept the file if it is otherwise valid.
The row numbers indicated by error messages take the instruction row into account.

Jira-Issue: MAV-1741