Create `SessionProgrammeYearGroup`
This will be used to store which year groups are being given to which
programmes in a particular session. Initially this will be the same as
the `Location::ProgrammeYearGroup` values for that particular location
to match the current logic, but this work will allow us to build
functionality where teams can focus on specific year groups in each
session.

Jira-Issue: MAV-2584