Switch to `SessionProgrammeYearGroup`
This converts the logic relating to which patients are shown in each
sessions to use the new `SessionProgrammeYearGroup` model introduced
in an earlier commit.

For now this will have no impact on the functionality for users, but
this change allows us to support more complex sessions in the future,
for example an HPV catch up session without year 8s.

Jira-Issue: MAV-2584