Add MMR second dose notification email template
This change introduces a new email template that replaces the standard
'We've decided your child is safe to vaccinate' message for MMR
patients eligible for their second dose. The new template is
automatically triggered when a child is moved out of 'Needs triage'
status and is eligible for an additional MMR dose.