Create `TeamLocation`
This is a new model which will be used to track which locations are
linked to which teams in each academic year. This is needed because the
current model means that when a location moves to a different team, any
views from previous academic years wouldn't be able to display that
location was once under the responsibility of a different team.

Jira-Issue: MAV-2584