Introduce `consent_not_given` outcome
Introduce a new outcome; `consent_not_given`. This is so we can
reach feature parity with the existing NIVS implementation. It's likely
that in the future this will be removed, as/when we rewrite the bulk
upload spec.

Add validation constraint that `outcome: consent_not_given`
can only apply to vaccination records whose source is `bulk_upload`.

Jira-Issue: MAV-