fix: school moves not updating after consent form gets matched
A consent form match can in some cases cause a change in the number of 
unresolved school moves, this was causing a bug where the school moves count 
displayed in the navigation bar diverged wildly from the actual count. 

This change resets the cache so that we recalculate the cache and keep the 
number accurate.

Jira-Issue: MAV-2636