Improve searching patients by name
# Filtering for which results to include
Previously, searches weren't matching on partial names, eg "jo" wouldn't match 
with "Johnathan Smith", because the similarity was too low. The new search will 
always include patients whose first or last name starts with the searched 
string.

In addition, when the search term is three characters or longer, any child 
whose name contains the search string will be returned. This catches children
with multiple names when searching for a name that is not their first given 
name. E.g. "Brad Pitt" should still find "PITT, William Bradley". Another way
to obtain that same behaviour would be to search by preferred name but this is
expensive as that field is both not indexed and encrypted.

When the search string contains multiple parts a child must match all parts in 
order to be shown in the results e.g. "Mariette Beau", will not match to
"Mariette Yearkey" nor will it match "Beau Adams".

# Ordering of the results
Previously, the search results were ordered based on a similarity comparison 
between the concatenated family and given name and the full query. 

Instead we take the sum of the strict word similarity to both given name and 
family name. Strict word similarity performs the same computation as similarity 
except the LHS is not compared to the full RHS but rather to any number of 
consecutive words in the RHS. This means that "Price Burt" compared to their 
family name "Burt" would only have similarity 0.54, while he would have a 
strict word similarity of 1.0.

# Performance of this new search
The new search is more performant than the old search (for now I have used 
EXPLAIN ANALYZE to assess the performance of the queries, but a performance
test will be ran as well before this gets included in production).

Improved performance is achieved by use of the trigram indices. Previously,
two CONCATs were used for each row in the database. In addition, WHERE clauses
can only use indices on boolean statements, so it was not possible to use the
trigram index on the SIMILARITY method. The 'left % right' operator performs 
the same check SIMILARITY(left, right) > pg_trgm.similarity_threshold, where 
pg_trgm.similarity_threshold defaults to 0.3, but could be set to a different
value if desired. [1]

[1]: https://www.postgresql.org/docs/18/pgtrgm.html


Jira-Issue: MAV-1093